+++ src/service/notificationApi/type.ts
@@ -1,0 +1,6 @@
+import { Notification } from "@contexts/AuthProvider/types";
+
+export interface ResGetNotifications {
+  contents: Notification[];
+  lastId: number | null;
+}
